Move trace link eval files inside integration package
