+++ tests/src/test/java/edu/kit/kastel/mcse/ardoco/core/tests/integration/tracelinks/eval/TestLink.java
@@ -1,5 +1,5 @@
 /* Licensed under MIT 2022. */
-package edu.kit.kastel.mcse.ardoco.core.tests.tracelinks.eval;
+package edu.kit.kastel.mcse.ardoco.core.tests.integration.tracelinks.eval;
 
 import edu.kit.kastel.mcse.ardoco.core.connectiongenerator.TraceLink;
