+++ tests/src/test/java/edu/kit/kastel/mcse/ardoco/core/tests/integration/tracelinks/eval/files/TLModelFile.java
@@ -1,4 +1,4 @@
-package edu.kit.kastel.mcse.ardoco.core.tests.tracelinks.eval.files;
+package edu.kit.kastel.mcse.ardoco.core.tests.integration.tracelinks.eval.files;
 
 import edu.kit.kastel.mcse.ardoco.core.common.AgentDatastructure;
 import edu.kit.kastel.mcse.ardoco.core.model.IModelInstance;
